Move display redaction into safety crate
