+++ crates/ironclaw_safety/Cargo.toml
@@ -19,6 +19,7 @@ serde_json = "1"
 thiserror = "2"
 tracing = "0.1"
 url = "2"
+urlencoding = "2"
 
 [dev-dependencies]
 criterion = "0.5"
